refactor(web): ♻️⚡️ extract FilterListsTable w/Loading state
