+++ web/app/loading.tsx
@@ -1,5 +1,5 @@
 import { FilterListsTable } from "@/components/filterlists-table/filterlists-table";
 
-export default function Home() {
-  return <FilterListsTable />;
+export default function Loading() {
+  return <FilterListsTable isLoading={true} />;
 }
