+++ web/components/filterlists-table/filterlist.ts
@@ -1,0 +1,11 @@
+export type FilterList = {
+  id: number;
+  name: string;
+  description?: string | null;
+  licenseId: number;
+  syntaxIds: number[];
+  languageIds: number[];
+  tagIds: number[];
+  primaryViewUrl?: string | null;
+  maintainerIds: number[];
+};
